Reorganized JAX link folder structure
